backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0
openai>=1.0
